nest_backend/azure_plant_service/requirements.txt
fastapi
uvicorn
python-multipart
transformers
torch 
torchvision
Pillow
